Resolve eslint errors
