+++ .eslintrc.json
@@ -14,7 +14,6 @@
     "no-shadow": "off",
     "no-param-reassign": "off",
     "no-plusplus": "off",
-    "no-cycle": "off",
     "eol-last": "off",
     "import/extensions": [ 1, {
       "js": "always", "json": "always"
